Centralize aggregate allowlist cap lookup in the action registry.
resolved_aggregate_cap and admission_cap_for_action were identical; keep one
canonical helper next to aggregate_allowlist_key and remove the duplicates.
